% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{narrow_dataset}
\alias{narrow_dataset}
\title{Compressibilty factor}
\format{A data frame with 57060 rows and 4 variables:
\describe{
  \item{Tpr}{price, in US dollars}
  \item{Ppr}{weight of the diamond, in carats}
  \item{z}{weight of the diamond, in carats}
  \item{set}{a set of Tpr curve}
  ...
}}
\usage{
narrow_dataset
}
\description{
A dataset containing the z factor from SK chart
}
\keyword{datasets}
