% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{sk_tidy_hp}
\alias{sk_tidy_hp}
\title{Compressibilty factor for High Pressure data points}
\format{A data frame with 904 rows and 4 variables:
\describe{
  \item{Tpr}{pseudo reduce temperature}
  \item{Ppr}{pseudo reduced pressure}
  \item{z}{compressibility factor}
  \item{set}{a set of Tpr curve}
  ...
}}
\usage{
sk_tidy_hp
}
\description{
A dataset containing the z factor from SK chart fow high Ppr
}
\keyword{datasets}
