% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{untidy}
\alias{untidy}
\title{untidy dataset}
\format{A data frame with 2853 rows and 60 variables:}
\usage{
untidy
}
\description{
The raw dataset read into R with first three column names changed
}
\keyword{datasets}
